reorganise types
