+++ src/connect/order/Customer.ts
@@ -1,26 +1,28 @@
 export type Customer = {
-    email?: string;
-    firstName?: string;
-    lastName?: string;
+    email?: string | null;
+    firstName?: string | null;
+    lastName?: string | null;
     /**
      * Should be formatted according to YYYY-MM-DD
      *
      * @example 1999-12-30
      */
-    birthDate?: string;
-    gender?: string;
-    phone?: string;
-    language?: string;
+    birthDate?: string | null;
+    gender?: string | null;
+    phone?: string | null;
+    language?: string | null;
     /** Locale code like "en_US" or "nl_BE" */
-    locale?: `${Lowercase<string>}_${Uppercase<string>}`;
-    ipAddress?: string;
+    locale?: `${Lowercase<string>}_${Uppercase<string>}` | null;
+    ipAddress?: string | null;
     trust?: number;
-    reference?: string;
-    gaClientId?: string;
-    company?: {
-        name?: string;
-        cocNumber?: string;
-        vatNumber?: string;
-        country?: string;
-    };
+    reference?: string | null;
+    gaClientId?: string | null;
+    company?: Company;
+};
+
+export type Company = {
+    name?: string | null;
+    cocNumber?: string | null;
+    vatNumber?: string | null;
+    country?: string | null;
 };
