--- src/connect/order/Notification.ts
@@ -1,4 +1,0 @@
-export type Notification = {
-    type?: 'push' | 'email';
-    recipient?: string;
-};
